add user rank calculation
