+++ src/components/film-details/film-details.js
@@ -1,0 +1,26 @@
+
+import {createFilmDetailsTemplate} from "./film-details-tpl";
+import {createElement} from "../../utils";
+
+export default class FilmDetails {
+  constructor(film) {
+    this._film = film;
+    this._element = null;
+  }
+
+  getTemplate() {
+    return createFilmDetailsTemplate(this._film);
+  }
+
+  getElement() {
+    if (!this._element) {
+      this._element = createElement(this.getTemplate());
+    }
+
+    return this._element;
+  }
+
+  removeElement() {
+    this._element = null;
+  }
+}
