+++ src/components/site-menu.js
@@ -1,22 +1,4 @@
-import {createElement} from "../utils";
-
-const calculateFilmsCountByFilter = (films, filterName) => {
-  let filmsCount = 0;
-
-  switch (filterName) {
-    case `watchlist`:
-      filmsCount = films.reduce((acc, film) => acc + (film.isInWatchlist ? 1 : 0), 0);
-      break;
-    case `history`:
-      filmsCount = films.reduce((acc, film) => acc + (film.isWatched ? 1 : 0), 0);
-      break;
-    case `favorites`:
-      filmsCount = films.reduce((acc, film) => acc + (film.isFavorite ? 1 : 0), 0);
-      break;
-  }
-
-  return filmsCount;
-};
+import {calculateFilmsCountByFilter, createElement} from "../utils";
 
 const createSiteMenuTemplate = (films) => {
   return (
