+++ src/components/user-rank.js
@@ -1,21 +1,33 @@
-import {createElement} from "../utils";
+import {calculateFilmsCountByFilter, createElement} from "../utils";
 
-export const createUserRankTemplate = () => {
+export const createUserRankTemplate = (films) => {
+  const watchedFilms = calculateFilmsCountByFilter(films, `history`);
+  let userRank = ``;
+
+  if (watchedFilms > 20) {
+    userRank = `Movie Buff`;
+  } else if (watchedFilms > 10) {
+    userRank = `Fan`;
+  } else if (watchedFilms > 0) {
+    userRank = `Novice`;
+  }
+
   return (
     `<section class="header__profile profile">
-      <p class="profile__rating">Movie Buff</p>
+      <p class="profile__rating">${userRank}</p>
       <img class="profile__avatar" src="images/bitmap@2x.png" alt="Avatar" width="35" height="35">
     </section>`
   );
 };
 
 export default class UserRank {
-  constructor() {
+  constructor(films) {
+    this._films = films;
     this._element = null;
   }
 
   getTemplate() {
-    return createUserRankTemplate();
+    return createUserRankTemplate(this._films);
   }
 
   getElement() {
