+++ src/main.js
@@ -5,13 +5,13 @@ import FilmsListsComponent from "./components/films-list";
 import FilmCardComponent from "./components/film-card";
 import NoFilmsComponent from "./components/no-films";
 import ShowMoreButtonComponent from "./components/show-more-button";
-import FilmDetailsComponent from "./components/film-details";
+import FilmDetailsComponent from "./components/film-details/film-details";
 import FilmsCountComponent from "./components/films-count";
 import {generateFilms} from "./mock/film";
 import {render, RenderPosition} from "./utils";
 
 const FilmSettings = {
-  TOTAL_COUNT: 18,
+  TOTAL_COUNT: 22,
   SHOWING_ON_START_COUNT: 5,
   SHOWING_BY_BUTTON: 5,
   EXTRA_COUNT: 2
@@ -107,7 +107,7 @@ const mostCommentedFilms = films
   .sort((a, b) => b.commentsCount - a.commentsCount)
   .slice(0, FilmSettings.EXTRA_COUNT);
 
-render(siteHeaderElement, new UserRankComponent().getElement(), RenderPosition.BEFOREEND);
+render(siteHeaderElement, new UserRankComponent(films).getElement(), RenderPosition.BEFOREEND);
 render(siteMainElement, new SiteMenuComponent(films).getElement(), RenderPosition.BEFOREEND);
 render(siteMainElement, new FilterComponent().getElement(), RenderPosition.BEFOREEND);
